pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yolo-cloud-training"
version = "0.1.0"
description = "Official Ultralytics YOLO11/YOLO26 orchestration for Vast.ai and Alibaba Cloud OSS."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "Jinja2>=3.1.0",
  "oss2>=2.18.5",
  "Pillow>=10.0.0",
  "PyYAML>=6.0",
  "ultralytics>=8.3.0",
  "uvicorn>=0.30.0",
  "vastai-sdk>=0.5.1",
]

[project.optional-dependencies]
dev = ["httpx>=0.27.0", "pytest>=8.0"]
pico = ["pycocotools>=2.0.7"]

[project.scripts]
yoloctl = "yoloctl.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"yoloctl" = ["review_static/*", "review_templates/*"]
